requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf
google-cloud-aiplatform>=1.38
langdetect
argostranslate
transformers
indic-nlp-library
torch
deep-translator
pymupdf
google-generativeai
sentence-transformers
matplotlib 
seaborn
plotly
